:recycle: refactor: Atualizacao da Função do Model ode Difusão
Alterado como a funçã oestava atribuída para o calculo da difusão das características. Adotado  o método sugerido pelo professor
